+++ src/Google.Analytics.SDK.Core/Google.Analytics.SDK.Core.csproj
@@ -25,10 +25,6 @@ Supported Platforms:
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="Logging\" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Logging" Version="2.1.0" />
     <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
   </ItemGroup>
