+++ src/Google.Analytics.SDK.Core/Logging/NoLogging.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using Microsoft.Extensions.Logging;
+
+namespace Google.Analytics.SDK.Core
+{
+    public class NoLogging : ILogger {
+        public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
+        {
+          
+        }
+
+        public bool IsEnabled(LogLevel logLevel)
+        {
+            return true;
+        }
+
+        public IDisposable BeginScope<TState>(TState state)
+        {
+            return null;
+        }
+    }
+}
