+++ tests/Google.Analytics.SDK.Tests/Trackers/MobileTrackerTests.cs
@@ -1,5 +1,6 @@
 using System;
 using Google.Analytics.SDK.Core;
+using Google.Analytics.SDK.Core.Extensions;
 using Google.Analytics.SDK.Core.Helper;
 using Google.Analytics.SDK.Core.Hits;
 using Google.Analytics.SDK.Core.Hits.MobileHits;
