+++ tests/Google.Analytics.SDK.Tests/Trackers/MockHit.cs
@@ -1,0 +1,12 @@
+using Google.Analytics.SDK.Core.Hits;
+
+namespace Google.Analytics.SDK.Tests.Trackers
+{
+    public class MockHit : HitBase
+    {
+        public MockHit()
+        {
+
+        }
+    }
+}
